+++ app/shared/io/helpers/AdapterHelper.scala
@@ -1,6 +1,7 @@
 package shared.io.helpers
 
 import shared.com.ortb.adapters.traits.implementations.BasicAdapterImplementation
+import shared.io.helpers.traits.AdapterHelperProperties
 
 object AdapterHelper
   extends BasicAdapterImplementation
