+++ app/shared/io/helpers/traits/AdapterHelperProperties.scala
@@ -1,4 +1,4 @@
-package shared.io.helpers
+package shared.io.helpers.traits
 
 import shared.com.ortb.adapters.traits.{ ModelsAdapters, ModelsAdaptersConcreteImplementation }
 import shared.com.ortb.adapters.{ EntityWrapperAdapterConcreteImplementation, RepositoryOperationResultModelAdapterConcreteImplementation }
